+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Streaming API to encrypt with compression:
+  - `func (keyRing *KeyRing) EncryptStreamWithCompression`
+  - `func (keyRing *KeyRing) EncryptSplitStreamWithCompression`
+  - `func (sk *SessionKey) EncryptStreamWithCompression`
+
 ## [2.5.0] 2022-12-16
 ### Changed
 - Update `github.com/ProtonMail/go-crypto` to the latest version